# docs/images/captures/openapi.json
{"openapi":"3.1.0","info":{"title":"Quantum Shield Core API","description":"Enterprise post-quantum cryptographic enclave.\n\n- **Algorithm**: ML-KEM-768 (Kyber768) + AES-256-GCM\n- **Audit**: HMAC-SHA256 signed logs with key rotation\n- **Auth**: API key (SHA-256 hashed) with RBAC\n\nAuthenticate with header `X-API-Key`.","version":"1.0.0"},"paths":{"/metrics":{"get":{"tags":["System"],"summary":"Metrics","description":"Endpoint that serves Prometheus metrics.","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["System"],"summary":"Liveness and database health","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/v1/keys/generate":{"post":{"tags":["Key Management"],"summary":"Generate ML-KEM-768 key pair","operationId":"generate_keys_api_v1_keys_generate_post","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KeyPairResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/crypto/seal":{"post":{"tags":["Cryptography"],"summary":"Hybrid encrypt (seal)","operationId":"seal_data_api_v1_crypto_seal_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SealRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SealResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/crypto/unseal":{"post":{"tags":["Cryptography"],"summary":"Hybrid decrypt (unseal)","operationId":"unseal_data_api_v1_crypto_unseal_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnsealRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnsealResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/audit/log":{"post":{"tags":["Audit Trail"],"summary":"Append signed audit entry","operationId":"write_audit_log_api_v1_audit_log_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/audit/logs":{"get":{"tags":["Audit Trail"],"summary":"List audit logs (newest first)","operationId":"read_audit_logs_api_v1_audit_logs_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"action","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Action"}},{"name":"actor","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Actor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditLogEntryResponse"},"title":"Response Read Audit Logs Api V1 Audit Logs Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/logs/{log_id}":{"get":{"tags":["Audit Trail"],"summary":"Get audit log by ID","operationId":"read_audit_log_api_v1_audit_logs__log_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"log_id","in":"path","required":true,"schema":{"type":"integer","title":"Log Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditLogEntryResponse"}}}},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/stats":{"get":{"tags":["Audit Trail"],"summary":"Audit log statistics","operationId":"get_audit_stats_api_v1_audit_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}}},"components":{"schemas":{"AuditLogEntryResponse":{"properties":{"id":{"type":"integer","title":"Id"},"action":{"type":"string","title":"Action"},"target":{"type":"string","title":"Target"},"actor":{"type":"string","title":"Actor"},"log_json":{"type":"string","title":"Log Json"},"signature":{"type":"string","title":"Signature"},"integrity":{"type":"string","title":"Integrity"}},"type":"object","required":["id","action","target","actor","log_json","signature","integrity"],"title":"AuditLogEntryResponse"},"AuditRequest":{"properties":{"action":{"type":"string","maxLength":64,"minLength":1,"title":"Action"},"target":{"type":"string","maxLength":256,"minLength":1,"title":"Target"},"user":{"type":"string","maxLength":64,"minLength":1,"title":"User"}},"type":"object","required":["action","target","user"],"title":"AuditRequest","example":{"action":"EXPORT","target":"report.pdf","user":"alice"}},"ErrorResponse":{"properties":{"detail":{"type":"string","title":"Detail"}},"type":"object","required":["detail"],"title":"ErrorResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"algorithm":{"type":"string","title":"Algorithm"},"version":{"type":"string","title":"Version"},"database":{"type":"string","title":"Database"}},"type":"object","required":["status","algorithm","version","database"],"title":"HealthResponse"},"KeyPairResponse":{"properties":{"public_key_b64":{"type":"string","title":"Public Key B64","description":"Kyber768 public key (Base64)."},"private_key_b64":{"type":"string","title":"Private Key B64","description":"Kyber768 secret key (Base64). Store securely."}},"type":"object","required":["public_key_b64","private_key_b64"],"title":"KeyPairResponse","example":{"private_key_b64":"def456...","public_key_b64":"abc123..."}},"SealRequest":{"properties":{"public_key_b64":{"type":"string","title":"Public Key B64","description":"Recipient Kyber768 public key."},"data_b64":{"type":"string","title":"Data B64","description":"Plaintext (Base64)."},"context":{"type":"string","maxLength":256,"minLength":1,"title":"Context"}},"type":"object","required":["public_key_b64","data_b64","context"],"title":"SealRequest","example":{"context":"contract-2025-001","data_b64":"SGVsbG8=","public_key_b64":"base64-public-key"}},"SealResponse":{"properties":{"ciphertext_pqc_b64":{"type":"string","title":"Ciphertext Pqc B64","description":"Kyber768 KEM ciphertext."},"nonce_b64":{"type":"string","title":"Nonce B64","description":"AES-GCM nonce (12 bytes)."},"encrypted_data_b64":{"type":"string","title":"Encrypted Data B64","description":"AES-256-GCM ciphertext + tag."}},"type":"object","required":["ciphertext_pqc_b64","nonce_b64","encrypted_data_b64"],"title":"SealResponse"},"UnsealRequest":{"properties":{"private_key_b64":{"type":"string","title":"Private Key B64"},"ciphertext_pqc_b64":{"type":"string","title":"Ciphertext Pqc B64"},"nonce_b64":{"type":"string","title":"Nonce B64"},"encrypted_data_b64":{"type":"string","title":"Encrypted Data B64"},"context":{"type":"string","maxLength":256,"minLength":1,"title":"Context"}},"type":"object","required":["private_key_b64","ciphertext_pqc_b64","nonce_b64","encrypted_data_b64","context"],"title":"UnsealRequest"},"UnsealResponse":{"properties":{"decrypted_data_b64":{"type":"string","title":"Decrypted Data B64","description":"Recovered plaintext (Base64)."}},"type":"object","required":["decrypted_data_b64"],"title":"UnsealResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"}}},"tags":[{"name":"System","description":"Health, metrics, and operational endpoints."},{"name":"Key Management","description":"Post-quantum key pair generation (ML-KEM-768)."},{"name":"Cryptography","description":"Hybrid seal/unseal operations (Kyber768 + AES-256-GCM)."},{"name":"Audit Trail","description":"HMAC-signed, append-only audit log with integrity verification."}]}